# src/main/gates/src/main/java/com/example/gates/model/Gate.java
package com.example.gates.model;

public class Gate {
    private int id;
    private String name;
    private String location;
    private String status; // "open" или "closed"

    // Конструкторы
    public Gate() {}

    public Gate(int id, String name, String location, String status) {
        this.id = id;
        this.name = name;
        this.location = location;
        this.status = status;
    }

    // Геттеры и сеттеры
    public int getId() { return id; }
    public void setId(int id) { this.id = id; }

    public String getName() { return name; }
    public void setName(String name) { this.name = name; }

    public String getLocation() { return location; }
    public void setLocation(String location) { this.location = location; }

    public String getStatus() { return status; }
    public void setStatus(String status) { this.status = status; }
}