# swagger.yaml
openapi: 3.0.0
info:
  title: Smart Home REST API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
paths:
...
  /gates:
    get:
      tags: [Gates]
      summary: Посмотреть список всех ворот
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gate'
    post:
      tags: [Gates]
      summary: Добавить новые ворота
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GateInput'
      responses:
        '201':
          description: Ворота добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
  /gates/{id}:
    get:
      tags: [Gates]
      summary: Найти ворота по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
        '404':
          description: Ворота не найдены
    put:
      tags: [Gates]
      summary: Открыть/закрыть ворота
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GateInput'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gate'
        '404':
          description: Ворота не найдены
    delete:
      tags: [Gates]
      summary: Удалить ворота
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Удалено успешно
        '404':
          description: Ворота не найдены
...
components:
  schemas:
...
    Gate:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Главные ворота"
        status:
          type: boolean
          example: false
    GateInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Главные ворота"
        status:
          type: boolean
          default: false
          example: false
...